[JSON-AST] Publishing native artifact on bintray
