+++ gradle.properties
@@ -17,6 +17,7 @@
 kotlin.version=1.2.50
 kotlin.version.snapshot=1.2-SNAPSHOT
 plugin.version=0.5.1
+konan.version=0.7
 
 org.gradle.parallel=true
 org.gradle.caching=true
