+++ json/common/build.gradle
@@ -1,0 +1,12 @@
+apply plugin: 'kotlin-platform-common'
+
+sourceSets {
+    main.kotlin.srcDirs = ['src']
+}
+
+dependencies {
+    compile libraries.kotlin_stdlib_common
+
+    testCompile libraries.kotlin_test_annotations_common
+    testCompile libraries.kotlin_test_common
+}
