+++ json/native/src/kotlinx/serialization/json/StringOps.kt
@@ -1,0 +1,7 @@
+package kotlinx.serialization.json
+
+actual fun CharArray.createString(length: Int): String =
+    StringBuilder().also {
+        it.insert(0, this)
+        it.length = length
+    }.toString()
