+++ runtime/common/build.gradle
@@ -18,6 +18,7 @@ apply plugin: 'kotlin-platform-common'
 
 dependencies {
     compile libraries.kotlin_stdlib_common
+    compile project(':jsonparser')
 
     testCompile libraries.kotlin_test_annotations_common
     testCompile libraries.kotlin_test_common
