--- runtime/common/src/main/kotlin/kotlinx/serialization/internal/CommonStrings.kt
@@ -1,25 +1,0 @@
-/*
- * Copyright 2017 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package kotlinx.serialization.internal
-
-/**
- * Creates a string by concatenating given chars.
- * Can be more efficient than `joinToString` on some platforms.
- *
- * charArrayOf('a','b','c').createString(2) = "ab"
- */
-expect fun CharArray.createString(length: Int): String
