+++ runtime/js/src/main/kotlin/kotlinx/serialization/json/StringOpsImpl.kt
@@ -1,0 +1,4 @@
+package kotlinx.serialization.json
+
+actual fun CharArray.createString(length: Int): String =
+    joinToString(separator = "", limit = length, truncated = "")
