+++ settings.gradle
@@ -19,11 +19,15 @@ rootProject.name = 'kotlinx-serialization-runtime'
 include ':common'
 include ':jvm'
 include ':js'
+include ':native'
+include ':jsonparser'
 include ':configparser'
 include ':benchmark'
 
 project(':common').projectDir = file('./runtime/common')
 project(':jvm').projectDir = file('./runtime/jvm')
 project(':js').projectDir = file('./runtime/js')
+project(':native').projectDir = file('./json/native')
+project(':jsonparser').projectDir = file('./json/common')
 project(':configparser').projectDir = file('./formats/config')
 project(':benchmark').projectDir = file('./benchmark')
